removed order back-end method

+++ TeamPizza/OrderApp/views.py
@@ -4,7 +4,7 @@ from django.conf import settings
 from django.core.handlers.wsgi import WSGIRequest
 from django.utils.functional import SimpleLazyObject
 from django.http.request import QueryDict
-from django.db import Error as DB_Error
+from django.db import DatabaseError as DB_Error
 
 import os
 import hashlib
@@ -157,11 +157,35 @@ def close_order(request: WSGIRequest, hash_id: str):
 
 
 @login_required(login_url='/login-required')
-def delete_order(request: WSGIRequest, hash_id: str):
+def delete_order(request: WSGIRequest):
+    logger: Logger = logging.getLogger(settings.LOGGER_NAME)
     user = request.user
     context = {'user': user}
-    # POST method to delete order in db
-    return render(request, 'OrderApp/order-deleted.html', context)
+    if request.method == "POST" and user.is_authenticated:
+        hash_id = ''
+        try:
+            post_body: QueryDict = request.POST
+            params: dict = post_body.dict()
+            hash_id = params['hash_id']
+        except KeyError as ke:
+            logger.error(f'Key error while user try remove order, probably someone change form!'
+                         f'  info: {ke.args}')
+        try:
+            order = Order.objects.filter(hash_id=hash_id).get()
+            if order and order.order_owner.id == user.id:
+                Order.objects.filter(hash_id=hash_id).delete()
+                return redirect('/operation-success/')
+        except DB_Error as db_err:
+            logger.error(f'Delete order by user {user.username} failed ! info: {db_err.args}')
+            context['bad_param'] = 'Something went wrong, try again'
+        except BaseException as be:
+            logger.error(f'Delete order by user {user.username} failed ! Base exception cached info: {be.args}')
+            context['bad_param'] = 'Something went wrong, try again'
+        return redirect('/operation-failed/')
+    elif not user.is_authenticated:
+        return render(request, 'TeamPizza/not-authenticated.html', context, status=401)
+    else:
+        return render(request, 'TeamPizza/bad-method.html', context, status=400)
 
 
 @login_required(login_url='/login-required')
@@ -189,10 +213,13 @@ def delete_contribution(request: WSGIRequest):
                 contribution = ContributionOrder.objects \
                     .filter(id=contribution_id, order=order).get()
                 logger.debug(f'User id is: {user.id} !')
-                if contribution and contribution.contribution_owner.id == user.id:
+                if contribution and contribution.contribution_owner.id == user.id and order.is_open:
                     ContributionOrder.objects.filter(id=contribution_id, order=order).delete()
                     logger.info(f'Delete order contribution by user {user.username} success !')
                     return redirect('/operation-success/')
+                elif not order.is_open:
+                    context['bad_param'] = f'Order has been closed at {order.close_time}'
+                    return render(request, 'TeamPizza/not-authenticated.html', context)
             except DB_Error as db_err:
                 logger.error(f'Delete order by user {user.username} failed ! info: {db_err.args}')
                 context['bad_param'] = 'Problem with database try again'
@@ -230,11 +257,14 @@ def update_contribution_get(request: WSGIRequest, hash_id: str, contribution_id:
         try:
             order = Order.objects.filter(hash_id=hash_id).get()
             contribution = ContributionOrder.objects.filter(id=contribution_id, order=order).get()
-            if contribution.contribution_owner.id == user.id:
+            if contribution.contribution_owner.id == user.id and order.is_open:
                 context['contribution'] = contribution
                 context['hash_id'] = hash_id
                 context['contribution_id'] = contribution_id
                 return render(request, 'OrderApp/contribution-update.html', context)
+            elif not order.is_open:
+                context['bad_param'] = f'Order has been closed at {order.close_time}'
+                return render(request, 'TeamPizza/not-authenticated.html', context)
             else:
                 return render(request, 'TeamPizza/not-authenticated.html', status=401)
         except DB_Error as db_err:
@@ -272,7 +302,7 @@ def update_contribution_post(request: WSGIRequest, hash_id: str, contribution_id
         try:
             order = Order.objects.filter(hash_id=hash_id).get()
             contribution = ContributionOrder.objects.filter(id=contribution_id, order=order).get()
-            if contribution.contribution_owner.id == user.id:
+            if contribution.contribution_owner.id == user.id and order.is_open:
                 pieces_int = 0
                 if pieces:
                     try:
@@ -294,6 +324,9 @@ def update_contribution_post(request: WSGIRequest, hash_id: str, contribution_id
                     context['bad_param'] = 'Bad params!'
                     context['contribution'] = contribution
                     return render(request, 'OrderApp/contribution-update.html', context)
+            elif not order.is_open:
+                context['bad_param'] = f'Order has been closed at {order.close_time}'
+                return render(request, 'TeamPizza/not-authenticated.html', context)
             else:
                 return render(request, 'TeamPizza/not-authenticated.html', status=401)
         except DB_Error as db_err:
